refactor: :memo: refactor using reactive from
remove old hexo state implementation
