+++ lib/blocs/ble_sensor_controller/bloc/gateway_controller_bloc.dart
@@ -1,0 +1,69 @@
+import 'dart:async';
+
+import 'package:bloc/bloc.dart';
+import 'package:equatable/equatable.dart';
+import 'package:gateway/blocs/ble_sensor/ble_sensor.dart';
+import 'package:gateway/blocs/ble_sensor/bloc/ble_sensor_bloc.dart';
+import 'package:gateway/blocs/ble_sensor_controller/controller_state.dart';
+import 'package:gateway/blocs/ble_sensor_controller/controller_state_from.dart';
+
+part 'gateway_controller_bloc_event.dart';
+part 'gateway_controller_bloc_state.dart';
+
+// TODO detect does having person in front
+class GatewayControllerBloc
+    extends Bloc<GatewayControllerEvent, GatewayControllerState> {
+  final List<BleSensorBloc> listOfSensor;
+  late StreamController<BleSensorOnData> streamBleSensorDataConsumer;
+
+  final HexoStateForm hexoState = HexoStateForm();
+
+  GatewayControllerBloc(this.listOfSensor)
+      : super(BleSensorControllerInitial()) {
+    streamBleSensorDataConsumer = StreamController();
+    _initListenSenorData();
+
+    // Send command to BLE sensor
+    // inner bloc communication
+    on<BleSensorControllerSensorCommandEvent>((event, emit) => {
+          listOfSensor
+              .firstWhere((sensor) => event.sensorType == sensor.sensorType)
+              .add(BleSensorWriteEvent(event.data))
+        });
+
+    // listen and add data to hexo State
+
+    on<GatewayControllerBLESensorDataEvent>(
+        (event, emit) => {hexoState.addBLEDataToForm(event)});
+
+    on<BleSensorControllerResetEvent>((event, emit) => {hexoState.resetFrom()});
+  }
+
+  _initListenSenorData() {
+    streamBleSensorDataConsumer.stream.listen((sensor) {
+      add(GatewayControllerBLESensorDataEvent(sensor.data, sensor.sensorType));
+    });
+    for (var sensor in listOfSensor) {
+      sensor.stream.where((event) => event is BleSensorOnData).listen((sensor) {
+        sensor = sensor as BleSensorOnData;
+        add(GatewayControllerBLESensorDataEvent(
+            sensor.data, sensor.sensorType));
+      });
+      // .pipe(streamBleSensorDataConsumer);
+    }
+  }
+
+  @override
+  void onChange(Change<GatewayControllerState> change) {
+    super.onChange(change);
+    print(change);
+  }
+
+  @override
+  Future<void> close() {
+    // streamHexoStateSub.cancel();
+    streamBleSensorDataConsumer.close();
+    // TODO: implement close
+    return super.close();
+  }
+}
